+++ OsiriXClasses/DICOMFiles/dicomFile.mm
@@ -210,11 +210,13 @@ char* replaceBadCharacter (char* str, NSStringEncoding encoding)
     [mutable1 replaceOccurrencesOfString:@"   " withString:@" " options:0 range:mutable1.range]; //tripple space -> single space
 	[mutable1 replaceOccurrencesOfString:@"  " withString:@" " options:0 range:mutable1.range]; //double space -> single space
     
-	int i = [mutable1 length];
+	NSUInteger i = [mutable1 length];
 	while( --i > 0)
 	{
-		if( [mutable1 characterAtIndex: i]==' ') [mutable1 deleteCharactersInRange: NSMakeRange( i, 1)];
-		else i = 0;
+		if ( [mutable1 characterAtIndex: i]==' ')
+            [mutable1 deleteCharactersInRange: NSMakeRange( i, 1)];
+		else
+            i = 0;
 	}
 	
 	return mutable1;
